fixed poducts not comming for some users

+++ lib/src/pages/report_table.dart
@@ -492,7 +492,7 @@ class _ReportTableState extends State<ReportTable> {
   showStockRow() {
     int productQTYTotal = 0;
     int salesValueTotal = 0;
-    int purchaseValueTotal = 0;
+    double purchaseValueTotal = 0;
     int marginValueTotal = 0;
     List<DataRow> list = [];
     for (int index = 0; index < widget.args.products!.length; index++) {
